+++ sdk/resourcemanager/databox/armdatabox/zz_generated_management_client.go
@@ -38,7 +38,7 @@ func NewManagementClient(subscriptionID string, credential azcore.TokenCredentia
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
-	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	ep := cloud.AzurePublic.Services[cloud.ResourceManager].Endpoint
 	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
 		ep = c.Endpoint
 	}
@@ -56,6 +56,7 @@ func NewManagementClient(subscriptionID string, credential azcore.TokenCredentia
 
 // Mitigate - Request to mitigate for a given job
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2022-02-01
 // jobName - The name of the job Resource within the specified resource group. job names must be between 3 and 24 characters
 // in length and use any alphanumeric and underscore only
 // resourceGroupName - The Resource Group Name
@@ -96,8 +97,8 @@ func (client *ManagementClient) mitigateCreateRequest(ctx context.Context, jobNa
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-12-01")
+	reqQP.Set("api-version", "2022-02-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, mitigateJobRequest)
 }